Abstract handling of loggers in tests.
